Fixed a segfault caused by integer overflow. See
https://github.com/lh3/bwa/commit/7d2c806e7f860ec03dfbf964075c36e73c81a0fb#commitcomment-11819978

+++ bwatpx.h
@@ -27,7 +27,7 @@
 #define b128_hash(a) ((uint32_t)(a).x)
 
 typedef struct {
-        int n;
+        int64_t n;
         bwtint_t *a;
 } poslist_t;
